pyproject.toml
[project]
name = "databridge"
version = "0.1.0"
description = "A simple configuration-driven ETL pipeline API built with Django and DRF."
requires-python = ">=3.13"
dependencies = [
    "django>=6.0.2",
    "djangorestframework>=3.16.1",
    "openpyxl>=3.1.5",
    "pandas>=3.0.1",
]
